Fastmidiin (#116)
Co-authored-by: Huw Cheston <hwc31@cam.ac.uk>
Co-authored-by: Mark Gotham <19486660+MarkGotham@users.noreply.github.com>

This is still not perfect, but I think a big improvement and much faster MIDI file input. Main todo items are to make the generated documentation give a good introduction to data structure and how to use it. More testing is needed. I'm seeing errors with tests and documentation around melsim, but I'm assuming that's unrelated.